feat: add firebase init singleton and upgrade junit to junit5
