+++ src/main/java/relay/data_access/FirestoreSingleton.java
@@ -4,39 +4,40 @@ import com.google.cloud.firestore.Firestore;
 import com.google.firebase.cloud.FirestoreClient;
 
 /**
- * This class provides a threadsafe singleton instance of Firestore database for interacting with Firebase Firestore.
+ * This class provides a threadsafe singleton instance of Firestore database for
+ * interacting with Firebase Firestore.
  */
 public class FirestoreSingleton {
 
-    // Singleton instance of FirestoreDatabase
-    private static volatile FirestoreSingleton singleton;
+	// Singleton instance of FirestoreSingleton
+	private static volatile FirestoreSingleton singleton;
 
-    // Reference to the Firestore database
-    private final Firestore databaseReference;
+	// Reference to the Firestore database
+	private final Firestore databaseReference;
 
-    /**
-     * Private constructor to initialize the Firestore database.
-     * The private constructor is necessary to make it a singleton.
-     */
-    private FirestoreSingleton() {
-        databaseReference = FirestoreClient.getFirestore();
-    }
+	/**
+	 * Private constructor to retrieve a single Firestore client instance.
+	 * The private constructor is necessary to make it a singleton.
+	 */
+	private FirestoreSingleton() {
+		databaseReference = FirestoreClient.getFirestore();
+	}
 
-    /**
-     * Returns the singleton's reference to FirestoreDatabase.
-     * This method is threadsafe.
-     *
-     * @return The Firestore database instance.
-     */
-    public static Firestore get() {
-        if (singleton == null) {
-            synchronized (FirestoreSingleton.class) {
-                // Recheck within the synchronized block
-                if (singleton == null) {
-                    singleton = new FirestoreSingleton();
-                }
-            }
-        }
-        return singleton.databaseReference;
-    }
+	/**
+	 * Returns the singleton's reference to FirestoreDatabase.
+	 * This method is threadsafe.
+	 *
+	 * @return The Firestore database instance.
+	 */
+	public static Firestore get() {
+		if (singleton == null) {
+			synchronized (FirestoreSingleton.class) {
+				// Recheck within the synchronized block
+				if (singleton == null) {
+					singleton = new FirestoreSingleton();
+				}
+			}
+		}
+		return singleton.databaseReference;
+	}
 }
