+++ src/test/java/RelayApplicationTests.java
@@ -1,0 +1,7 @@
+import org.springframework.boot.test.context.SpringBootTest;
+import relay.RelayApplication;
+
+@SpringBootTest(classes = RelayApplication.class)
+class RelayApplicationTests {
+
+}
