--- src/test/java/relay/RelayApplicationTests.java
@@ -1,13 +1,0 @@
-package relay;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class RelayApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
